[Core]refactor aqlm quant ops  (#4351)
